Add command to lookup trusted task data
This is pretty handy for debugging or troubleshooting trusted task
problems. If you provide a string as the param it will show just
tasks with names that include that string.

I don't want this script to become a "do everything" utility, but
here we're essentially just adding a command to access functionality
that already exists in the script, so I think it's fine.