Better handling of missing dependencies
I had a user who had the python/pip version of yq, and didn't have
oras. The errors produced in that situation are hard to diagnose, so
let's handle that better.

Actually some commands don't require all the dependencies. I was
initially going to split it up per command, but I figured it's
better to keep it simple.