Automatically select oci-ta tasks if appropriate
If the pipeline looks like a trusted artifact pipeline, and if the
task specified has a matching "-oci-ta" version, automatically use
that "-oci-ta" version of the task.

This simplifiles the "one-liner" instructions since users no longer
have to figure out which of the two versions to use. The script will
decide automatically if it needs to use the trusted-artifact version
of a task.