Cache the oras blob fetch of the trusted task data
In an upcoming commit I want to use this multiple times in a single
command. The trusted data bundle is large, so I don't want to be
doing the oras blob fetch over and over unnecessarily. Use a
temporary file to cache the data.

(The in-memory memoization used by the list_known_tasks function
doesn't work when we're running in a subshell so that's why we're
doing it a different way.)