Use a custom git config in shellspec tests
Without this we have problems running in the GitHub workflow. It's
not strictly needed when running locally, but we might as well be
consistent.

Use it for the shellspec install also since there are git clones
happening in that script and the git config makes the GitHub
workflow output tidier.