Experimental migrate to build-image-index support
Because the sast tasks expect the build-image-index task to be
present, so let's try to provide some assistance adding it to a
pipeline.

(As mentioned in the help, this is experimental and not well tested.
I figure even if it doesn't work in all cases, it could still be
useful to apply the expect changes quickly rather than doing it by
hand. Anway we'll see how it goes.)

Ref: https://issues.redhat.com/browse/EC-1202