Experimental migrate to build-image-index support
Originally this was going to be the suggested way to add the new
sast tasks to older pipelines without the build-image-index task.

But the backwards compatibility sed replace workaround implemented
in #4 seemed like a simpler and better approach, so as far as I know
this hasn't been used in a real world scenario, and maybe it won't.

But still, it pretty much works, it has some test coverage, and it
could inspire or evolve into a useful pipeline migration tool. At
the very least it serves to illustrate the kind of things that a
pipeline migration tool might need to handle. So for these reasons
I'm going to merge the PR and keep this code.

Ref: https://issues.redhat.com/browse/EC-1202