Document optional before-task-name add-tasks arg
There's no solid reason not to document this from the beginnning,
but I was assuming the default would be fine for most users, and
that it wouldn't be needed much.

Since then I've seen several pipelines without the
'deprecated-image-check' task, so I figure it's better to mention
it in the command help.