Tweak expected output to get tests passing
The test suite pulls in the real pipeline task yaml, so it's
sensitive to changes in the pipelines. Update the tests accordingly
to get them passing again.

(Ideally we'd do something making the tests more stable, e.g. not
use the real production main branch build-definitions, but I don't
feel it's worth the effort right now.)