Make tkn-bundle task visible
Since it's not in any of the pipelines we currently inspect, add the
new pipelines that contain it.

I'm doing this because I need it myself since I want to update the
ec-cli build pipeline to use the 0.2 version of the tkn-bundle task,
and I want to use the task-yaml command make sure I have the right
params.

It's related to needing the org.opencontainers.image.revision
annotation for https://issues.redhat.com/browse/EC-913