Add a format-yaml command
As explained in the comments, if you have to make formatting
changes, it's better to not make them along with the significant
change. This is a quick/lazy command to make that easier.

Includes the addition of a new function to find yaml files without
using git, and without caring what's in them.